+++ AUDIO_ONLY_EMP/package.json
@@ -15,7 +15,7 @@
     "@types/prettier": "2.6.0",
     "@typescript-eslint/eslint-plugin": "^5.40.1",
     "@typescript-eslint/parser": "^5.40.1",
-    "aws-cdk": "2.87.0",
+    "aws-cdk": "2.1023.0",
     "cdk-nag": "2.21.50",
     "jest": "^27.5.1",
     "ts-jest": "^27.1.4",
@@ -23,7 +23,7 @@
     "typescript": "~3.9.7"
   },
   "dependencies": {
-    "aws-cdk-lib": "2.202.0",
+    "aws-cdk-lib": "2.208.0",
     "constructs": "^10.0.0",
     "source-map-support": "^0.5.21"
   }
