+++ EMX_SECRET_SHARING/emx-entitlement-consumer/package.json
@@ -18,7 +18,7 @@
     "@types/node": "10.17.27",
     "@typescript-eslint/eslint-plugin": "^5.40.1",
     "@typescript-eslint/parser": "^5.40.1",
-    "aws-cdk": "2.177.0",
+    "aws-cdk": "2.1023.0",
     "cdk-nag": "2.21.50",
     "eslint": "^8.31.0",
     "eslint-config-prettier": "^8.5.0",
@@ -30,7 +30,7 @@
     "typescript": "~3.9.7"
   },
   "dependencies": {
-    "aws-cdk-lib": "2.177.0",
+    "aws-cdk-lib": "2.208.0",
     "aws-sdk": "^2.1354.0",
     "constructs": "^10.0.0",
     "source-map-support": "^0.5.21"
