+++ LIVE2VOD/package.json
@@ -16,8 +16,8 @@
     "@types/prettier": "2.6.0",
     "@typescript-eslint/eslint-plugin": "^5.40.1",
     "@typescript-eslint/parser": "^5.40.1",
-    "aws-cdk": "2.98.0",
-    "aws-cdk-lib": "2.98.0",
+    "aws-cdk": "2.1023.0",
+    "aws-cdk-lib": "2.208.0",
     "cdk-nag": "2.21.50",
     "constructs": "^10.0.0",
     "jest": "^27.5.1",
