+++ LIVE_THUMBNAIL_MONITORING/package.json
@@ -17,8 +17,8 @@
     "@types/prettier": "2.6.0",
     "@typescript-eslint/eslint-plugin": "^5.40.1",
     "@typescript-eslint/parser": "^5.40.1",
-    "aws-cdk": "2.87.0",
-    "aws-cdk-lib": "2.87.0",
+    "aws-cdk": "2.1023.0",
+    "aws-cdk-lib": "2.208.0",
     "cdk-nag": "2.21.50",
     "constructs": "^10.0.0",
     "jest": "^27.5.1",
