+++ PRIVATE_LIVE/package.json
@@ -16,7 +16,7 @@
     "@types/prettier": "2.6.0",
     "@typescript-eslint/eslint-plugin": "^5.40.1",
     "@typescript-eslint/parser": "^5.40.1",
-    "aws-cdk": "2.87.0",
+    "aws-cdk": "2.1023.0",
     "cdk-nag": "2.21.50",
     "jest": "^27.5.1",
     "ts-jest": "^27.1.4",
@@ -24,7 +24,7 @@
     "typescript": "~3.9.7"
   },
   "dependencies": {
-    "aws-cdk-lib": "2.178.0",
+    "aws-cdk-lib": "2.208.0",
     "constructs": "^10.0.0",
     "source-map-support": "^0.5.21"
   }
