+++ SSAI_CMAF_AND_DASH/package.json
@@ -14,14 +14,14 @@
     "@types/jest": "^27.5.2",
     "@types/node": "10.17.27",
     "@types/prettier": "2.6.0",
-    "aws-cdk": "^2.87.0",
+    "aws-cdk": "2.1023.0",
     "jest": "^27.5.1",
     "ts-jest": "^27.1.4",
     "ts-node": "^10.9.1",
     "typescript": "~3.9.7"
   },
   "dependencies": {
-    "aws-cdk-lib": "^2.87.0",
+    "aws-cdk-lib": "2.208.0",
     "cdk-nag": "^2.20.10",
     "constructs": "^10.0.0",
     "source-map-support": "^0.5.21"
