+++ SUSTAINABILITY_LIVE/package.json
@@ -18,13 +18,13 @@
     "@typescript-eslint/parser": "^5.40.1",
     "jest": "^27.5.1",
     "ts-jest": "^27.1.4",
-    "aws-cdk": "2.87.0",
+    "aws-cdk": "2.1023.0",
     "cdk-nag": "2.21.50",
     "ts-node": "10.4.0",
     "typescript": "4.4.4"
   },
   "dependencies": {
-    "aws-cdk-lib": "2.87.0",
+    "aws-cdk-lib": "2.208.0",
     "constructs": "^10.0.0",
     "source-map-support": "^0.5.21"
   }
